fix(map): player gets shown after locationPermission
- changed behaviour so that Permission get checked before map is loaded
