+++ app/src/main/java/com/github/cstrerath/uncover/ui/screens/map/MapScreenController.kt
@@ -4,7 +4,16 @@ import android.content.Context
 import android.content.Intent
 import android.util.Log
 import androidx.activity.result.ActivityResultLauncher
+import androidx.compose.foundation.isSystemInDarkTheme
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.viewinterop.AndroidView
 import com.github.cstrerath.uncover.domain.map.managers.MapManager
 import com.github.cstrerath.uncover.domain.map.managers.QuestMarkerHandler
 import com.github.cstrerath.uncover.ui.factories.MapViewFactory
@@ -47,4 +56,21 @@ class MapScreenController(
             isDarkMode = isDarkMode
         )
     }
+}
+
+@Composable
+fun MapScreen(questLauncher: ActivityResultLauncher<Intent>) {
+    val scope = rememberCoroutineScope()
+    val context = LocalContext.current
+    val isDarkMode = isSystemInDarkTheme()
+    val controller = remember { MapScreenController(context, scope) }
+
+    LaunchedEffect(Unit) {
+        controller.refreshMapData(questLauncher)
+    }
+
+    AndroidView(
+        factory = { controller.createMapView(isDarkMode) },
+        modifier = Modifier.fillMaxSize()
+    )
 }
